requirements.txt
streamlit
requests
python-dotenv
huggingface_hub
streamlit-option-menu
streamlit-tree-select
